Switch from `System.in`/`System.out` to `Console`
The `main` method now uses the system console to read input and write
output, instead of relying on the `System.in` and `System.out` fields.
This improves testability of the `main` method. The console is provided
through a utility class in order to make it mockable/overwritable in the
unit test.